introduce `overPath` object function
`overPath` : [String] -> (a -> b) -> {k:v} -> {k:v}

Makes a shallow clone of an object, setting or overriding the nodes
required to create the given path, and applying the given function to
the tail end of that path. Note that this copies and flattens prototype
properties onto the new object as well. All non-primitive properties
are copied by reference.